On second thought, move Size() and Order() into atomic interfaces.
It feels a little icky to do this, but this pairs the size-counting
methods with the exact method that would otherwise be used to produce
the same data. If the iteration can happen, it is guaranteed the
counting can happen - and possibly much more efficiently than by going
through iteration.